Define cdn-data-row type in schema Extract CdnUsageTable class
