+++ packages/api/src/schema/schema.yaml
@@ -29,10 +29,49 @@ tags:
 
 components:
   schemas:
+    cdn-data-row:
+      type: object
+      description:
+        Fields names taken from log format description
+        https://support.highwinds.com/hc/en-us/articles/360052181571-Raw-Log-Access-through-GCS
+      additionalProperties: false
+      required:
+        - unique_users
+        - total_sc_bytes
+      # oneOf: # allows validation, but generates too generic type definition
+      #   - required:
+      #       - stream_id
+      #   - required:
+      #       - playback_id
+      properties:
+        stream_id:
+          type: string
+          description: Session UUID. Used in recordings URLs.
+        playback_id:
+          type: string
+          description: Playback ID.
+        unique_users:
+          type: number
+          description:
+            Number of unique IP addresses of clients. Not accurate because
+            StackPath zeroes last part of the IP address.
+        total_filesize:
+          type: number
+          description: Size of the asset being delivered.
+        total_cs_bytes:
+          type: number
+          description: The total size of the request header.
+        total_sc_bytes:
+          type: number
+          description: Total bytes in the response to the client.
+        count:
+          type: number
+          description: Number of parsed log lines
     cdn-data-payload:
       type: array
       items:
         type: object
+        additionalProperties: false
         required:
           - region
           - file_name
@@ -49,30 +88,7 @@ components:
             type: array
             minItems: 1
             items:
-              type: object
-              required:
-                - unique_users
-                - total_sc_bytes
-              oneOf:
-                - required:
-                    - stream_id
-                - required:
-                    - playback_id
-              properties:
-                stream_id:
-                  type: string
-                playback_id:
-                  type: string
-                unique_users:
-                  type: number
-                total_filesize:
-                  type: number
-                total_cs_bytes:
-                  type: number
-                total_sc_bytes:
-                  type: number
-                count:
-                  type: number
+              $ref: "#/components/schemas/cdn-data-row"
     cdn-usage-last:
       type: object
       table: cdn_usage_last
